+++ Homework_02_Med_BS_WithChallenge.vb
@@ -1,25 +1,25 @@
 Sub HomeworkMed()
 
 Dim ws As Worksheet
-    
+
 For Each ws In ThisWorkbook.Worksheets
     ws.Activate
 
 	Dim ticker As String
-	Dim totalvolume, OpenYear, CloseYear, YearlyChange, PercentChange, lrow As Double
+	Dim totalvolume, OpenYear, CloseYear, YearlyChange, PercentChange, lrow As Long
 	Dim output As Integer
-	
+
 
 		output = 2
 		Cells(1, 9) = "Ticker"
 		Cells(1, 10) = "Yearly Change"
 		Cells(1, 11) = "Percent Change"
 		Cells(1, 12) = "Total Volume"
-		
+
 		lrow = Cells(Rows.Count, 1).End(xlUp).Row
-		
+
 		OpenYear = Cells(2, 3).Value
-		
+
 		For i = 2 To lrow
 			ticker = Cells(i, 1)
 			If ticker <> Cells(i + 1, 1) Then
@@ -34,31 +34,28 @@ For Each ws In ThisWorkbook.Worksheets
 				Cells(output, 11).NumberFormat = "0.00%"
 				totalvolume = 0
 				OpenYear = Cells(i + 1, 3).Value
-					
+
 					If YearlyChange >= 0 Then
 					Cells(output, 10).Interior.ColorIndex = 4
-					
+
 					ElseIf YearlyChange < 0 Then
 					Cells(output, 10).Interior.ColorIndex = 3
-					
+
 					End If
-					
-				
-					
+
+
+
 				output = output + 1
-					
+
 			ElseIf ticker = Cells(i + 1, 1) Then
 				totalvolume = totalvolume + Cells(i, 7)
-			
+
 			ElseIf Cells(i, 1) = " " Then
 				Exit For
-			
+
 			End If
-		   
+
 		Next i
 	Next ws
 
 End Sub
-
-
-
